requirements.txt
viam-sdk==0.45.2

typing-extensions

numpy>=1.24.0
Pillow>=10.0.0
